+++ CHANGES.md
@@ -11,9 +11,19 @@
 
 ## develop
 
+- [UPDATE] Sora C++ SDK を `2025.4.0` に上げる
+  - それに伴って以下のライブラリのバージョンも上げる
+  - libwebrtc のバージョンを `m138.7204.0.1` に上げる
+    - `rtc::` を `webrtc::` に変更する
+    - `cricket::` を `webrtc::` に変更する
+    - `ZakuroAudioDeviceModuleConfig` のメンバー `TaskQueueFactory` を削除し、都度 webrtc::CreateEnvironment() で webrtc::Environment を生成するようにする
+    - `webrtc::AudioDeviceModule::Create` から `webrtc::CreateAudioDeviceModule` に変更する
+  - CMake のバージョンを `4.0.3` に上げる
+  - @miosakuma
 - [UPDATE] Blend2D を公式サイトからダウンロードするように変更する
   - 今までは現時点の master のコミットハッシュを使っていたが、asmjit と一緒にバージョンを管理しないといけなかった
   - 公式サイトからのダウンロードだと asmjit を内包してるのでこっちの方が管理が楽そうという判断
+  - @melpon
 - [ADD] signaling URL のバリデーションを追加
   - signaling URL が ws:// または wss:// で始まらない場合はエラーを出力
   - @voluntas